Fix PHPT tests
