+++ tests/phpt/error_handler_called.phpt
@@ -1,13 +1,12 @@
 --TEST--
-Test that the error handler is not called regardless of the current
-`error_reporting` setting if its own `captureAt` configuration doesn't match
-the level of the thrown error.
+Test that the error handler is called regardless of the current `error_reporting` setting
 --FILE--
 <?php
 
 namespace Sentry\Tests;
 
 use Sentry\ErrorHandler;
+use Sentry\Tests\Fixtures\classes\StubErrorListener;
 
 $vendor = __DIR__;
 
@@ -19,9 +18,9 @@ require $vendor . '/vendor/autoload.php';
 
 error_reporting(E_ERROR);
 
-ErrorHandler::register(function () {
+ErrorHandler::addErrorListener(new StubErrorListener(function () {
     echo 'Callback invoked' . PHP_EOL;
-});
+}));
 
 echo 'Triggering error' . PHP_EOL;
 trigger_error(E_WARNING);
