+++ tests/phpt/fatal_error.phpt
@@ -6,8 +6,9 @@ Test catching fatal errors
 namespace Sentry\Tests;
 
 use PHPUnit\Framework\Assert;
+use Sentry\ErrorHandler;
 use function Sentry\init;
-use Sentry\State\Hub;
+use Sentry\State\Hub;use Sentry\Tests\Fixtures\classes\StubErrorListener;
 
 $vendor = __DIR__;
 
@@ -22,17 +23,9 @@ init([
     'send_attempts' => 1,
 ]);
 
-register_shutdown_function('register_shutdown_function', function () {
-    $client = Hub::getCurrent()->getClient();
-
-    /** @var \Sentry\Transport\HttpTransport $transport */
-    $transport = Assert::getObjectAttribute($client, 'transport');
-
-    Assert::assertAttributeEmpty('pendingRequests', $transport);
-    Assert::assertNotNull(Hub::getCurrent()->getLastEventId());
-
-    echo 'Shutdown function called';
-});
+ErrorHandler::addErrorListener(new StubErrorListener(function () {
+    echo 'Listener called';
+}));
 
 class TestClass implements \Serializable
 {
@@ -40,4 +33,4 @@ class TestClass implements \Serializable
 ?>
 --EXPECTF--
 Fatal error: Class Sentry\Tests\TestClass contains 2 abstract methods and must therefore be declared abstract or implement the remaining methods (Serializable::serialize, Serializable::unserialize) in %s on line %d
-Shutdown function called
+Listener called
